requirements.txt
# LangChain and Graph Dependencies
langchain==0.1.12
langchain-core==0.1.32
langchain-anthropic==0.1.4
langchain-openai==0.0.8
langgraph==0.0.27
langsmith==0.1.22

# LLM Providers
anthropic==0.18.1
openai==1.14.1
ollama==0.1.6

# ROS2 Dependencies
rclpy>=0.1.0

# Computer Vision Dependencies
opencv-python==4.9.0.80
numpy>=1.24.0
ultralytics==8.1.28

# Utilities
python-dotenv==1.0.1
aiohttp==3.9.3
typing-extensions>=4.5.0
pydantic>=2.0.0

# ROS2 Vision Dependencies
cv-bridge==3.2.1
sensor-msgs-py==4.0.0
visualization-msgs==4.0.0
